Fix CodeQL security scanning issues in aie-codegen (#63)
* Setup Copyrights automation (#47)

* Create copyright_updates.yml

* Update copyright_updates.yml

* Create COPYRIGHT_EXCLUDES

* Create languages_config.yml

* Create README.md

* Update README.md

* updated the runner and branch correctly.

* Updated the PR reviewers.

* Updated multiple pr reviewers.

* Updated the pr reviewers.

* Updated pr_team_reviewers in workflow.

* removedthe pr_team_revieviewers key.

* src: Fix Init and Reset CDO mismatch (#46)

In Timer Sync function, no need to clear Broadcast
configuration for second broadcast channel as it
was not configured. So removed function to clear
Broadcast configuration for second channel.
Also For AIE1 Reset CDO was differ because of Isolation
was not happening for all tiles. Which has been fix in
this PR.


* Update copyright headers (#49)


* Fix CodeQL security scanning issues in aie-codegen

Fixed issues:
- #5464, #5465: Comparison of narrow type with wide type in loop condition (xaie_reset_aie.c:98,110)
- #5462, #5463: Comparison of narrow type with wide type in loop condition (xaie_io_privilege.c:436,474)
- #5457: Multiplication result converted to larger type (xaie_ss.c:1111)
- #5456: Multiplication result converted to larger type (xaie_locks_aie4.c:261)
- #5452, #5453, #5454, #5455: Multiplication result converted to larger type (xaie_dma_aieml.c)
- #5449, #5450, #5451: Multiplication result converted to larger type (xaie_dma_aie.c)

Changes include:
- Changed loop variables from u8 to u32 to match comparison type
- Added explicit casts to u64 before multiplication to prevent overflow

---------

+++ src/dma/xaie_dma_aie.c
@@ -1354,8 +1354,9 @@ AieRC _XAie_DmaGetPendingBdCount(XAie_DevInst *DevInst, XAie_LocType Loc,
 	u64 Addr;
 	u32 StatusReg, StartQSize, Stalled, Status;
 
+	/* Issue #5451: Cast to u64 before multiplication to prevent overflow */
 	Addr = XAie_GetTileAddr(DevInst, Loc.Row, Loc.Col) +
-		DmaMod->ChStatusBase +(u32)Dir * DmaMod->ChStatusOffset;
+		DmaMod->ChStatusBase + (u64)Dir * DmaMod->ChStatusOffset;
 
 	RC = XAie_Read32(DevInst, Addr, &StatusReg);
 	if(RC != XAIE_OK) {
@@ -1429,8 +1430,9 @@ AieRC _XAie_DmaWaitForDone(XAie_DevInst *DevInst, XAie_LocType Loc,
 	u32 Mask, Value;
 	AieRC Status = XAIE_OK;
 
+	/* Issue #5450: Cast to u64 before multiplication to prevent overflow */
 	Addr = XAie_GetTileAddr(DevInst, Loc.Row, Loc.Col) +
-		DmaMod->ChStatusBase + (u32)Dir * DmaMod->ChStatusOffset;
+		DmaMod->ChStatusBase + (u64)Dir * DmaMod->ChStatusOffset;
 	Mask = DmaMod->ChProp->DmaChStatus[ChNum].AieDmaChStatus.Status.Mask |
 		DmaMod->ChProp->DmaChStatus[ChNum].AieDmaChStatus.StartQSize.Mask |
 		DmaMod->ChProp->DmaChStatus[ChNum].AieDmaChStatus.Stalled.Mask;
@@ -1487,8 +1489,9 @@ AieRC _XAie_DmaGetChannelStatus(XAie_DevInst *DevInst, XAie_LocType Loc,
 	u32 Mask;
 	AieRC RC;
 
+	/* Issue #5449: Cast to u64 before multiplication to prevent overflow */
 	Addr = XAie_GetTileAddr(DevInst, Loc.Row, Loc.Col) +
-		DmaMod->ChStatusBase + (u32)Dir * DmaMod->ChStatusOffset;
+		DmaMod->ChStatusBase + (u64)Dir * DmaMod->ChStatusOffset;
 
 	Mask = DmaMod->ChProp->DmaChStatus[ChNum].AieDmaChStatus.Status.Mask |
 		DmaMod->ChProp->DmaChStatus[ChNum].AieDmaChStatus.StartQSize.Mask |
